requirements.txt
google-generativeai
openai
requests
python-dotenv
pytrends
schedule
pillow
python-telegram-bot
tweepy
facebook-sdk
